removed cyclic import, employees see their trips now
Added a LoC script
Employees can now look at their assigned trips and see who is on their crew

+++ Fahrplansystem/app/static/js/main.js
@@ -16,63 +16,63 @@ $(document).ready(function() {
     })(jQuery);
 
     //Delete a user on button click
-    $('.manage_employees .delete_button').click(function() {
+    $('.delete_employee').click(function() {
         if (!window.confirm("Sind Sie sich sicher?")) {
             return
         }
         var employee_id = $(this).attr('employee_id')
         var ajaxReq = $.ajax({
-                url: '/manage_employees/'+employee_id,
-                type: 'DELETE',
-                statusCode: {
-                    200: function() {
-                        window.location.reload()
-                    },
-                    500: function() {
-                        $("#deleteEmployeeFailed").temporaryRemoveClass("hidden", 3000);
-                    }
+            url: '/manage_employees/'+employee_id,
+            type: 'DELETE',
+            statusCode: {
+                200: function() {
+                    window.location.reload()
+                },
+                500: function() {
+                    $("#deleteEmployeeFailed").temporaryRemoveClass("hidden", 3000);
                 }
-            });
+            }
+        });
     });
 
     // Delete crew on button click
-    $('.manage_crews .delete_button').click(function() {
+    $('.delete_crew').click(function() {
         if (!window.confirm("Sind Sie sich sicher?")) {
             return
         }
         var tour_id = $(this).attr('crew_id')
         var ajaxReq = $.ajax({
-                url: '/manage_crews/'+tour_id,
-                type: 'DELETE',
-                statusCode: {
-                    200: function() {
-                        window.location.reload()
-                    },
-                    500: function() {
-                        $("#deleteCrewFailed").temporaryRemoveClass("hidden", 3000);
-                    }
+            url: '/manage_crews/'+tour_id,
+            type: 'DELETE',
+            statusCode: {
+                200: function() {
+                    window.location.reload()
+                },
+                500: function() {
+                    $("#deleteCrewFailed").temporaryRemoveClass("hidden", 3000);
                 }
-            });
+            }
+        });
     });
 
     //Delete tour on button click
-    $('.manage_tours .delete_button').click(function() {
+    $('.delete_tour').click(function() {
         if (!window.confirm("Sind Sie sich sicher?")) {
             return
         }
         var tour_id = $(this).attr('tour_id')
         var ajaxReq = $.ajax({
-                url: '/manage_tours/'+tour_id,
-                type: 'DELETE',
-                statusCode: {
-                    200: function() {
-                        window.location.reload()
-                    },
-                    500: function() {
-                        $("#deleteTourFailed").temporaryRemoveClass("hidden", 3000);
-                    }
+            url: '/manage_tours/'+tour_id,
+            type: 'DELETE',
+            statusCode: {
+                200: function() {
+                    window.location.reload()
+                },
+                500: function() {
+                    $("#deleteTourFailed").temporaryRemoveClass("hidden", 3000);
                 }
-            });
+            }
+        });
     });
 
      $('.delete_trip').click(function() {
@@ -81,26 +81,36 @@ $(document).ready(function() {
         }
         var trip_id = $(this).attr('trip_id')
         var ajaxReq = $.ajax({
-                url: '/manage_tours/'+tour_id,
-                type: 'DELETE',
-                statusCode: {
-                    200: function() {
-                        window.location.reload()
-                    },
-                    500: function() {
-                        $("#deleteTourFailed").temporaryRemoveClass("hidden", 3000);
-                    }
+            url: '/manage_trips/'+ trip_id,
+            type: 'DELETE',
+            statusCode: {
+                200: function() {
+                    window.location.reload()
+                },
+                500: function() {
+                    $("#deleteTripFailed").temporaryRemoveClass("hidden", 3000);
                 }
-            });
+            }
+        });
     });
 
-    // Add empty crew
-    $('.manage_crews .add_empty_crew').click(function() {
+    $('.delete_rushhour').click(function() {
+        if (!window.confirm("Sind Sie sich sicher?")) {
+            return
+        }
+        var rushhour_id = $(this).attr('rushhour_id')
         var ajaxReq = $.ajax({
-                url: '/add_empty_crew',
-                type: 'POST'
-            });
-        window.location.reload()
+            url: '/delete_rushhour/'+ rushhour_id,
+            type: 'DELETE',
+            statusCode: {
+                200: function() {
+                    window.location.reload()
+                },
+                500: function() {
+                    $("#deleteTripFailed").temporaryRemoveClass("hidden", 3000);
+                }
+            }
+        });
     });
 
 });
